refactor(numbers_table): move to storages/memory
