+++ query/src/storages/memory/numbers_table_stream.rs
