+++ query/src/table_functions/table_function_factory.rs
@@ -26,7 +26,7 @@ use crate::catalogs::SYS_TBL_FUC_ID_END;
 use crate::catalogs::SYS_TBL_FUNC_ID_BEGIN;
 use crate::storages::fuse::FuseHistoryTable;
 use crate::storages::fuse::FUSE_FUNC_HIST;
-use crate::table_functions::NumbersTable;
+use crate::storages::memory::NumbersTable;
 use crate::table_functions::TableFunction;
 
 pub type TableArgs = Option<Vec<Expression>>;
