+++ query/tests/it/main.rs
@@ -26,6 +26,5 @@ mod servers;
 mod sessions;
 mod sql;
 mod storages;
-mod table_functions;
 mod tests;
 mod users;
