+++ query/tests/it/storages/memory/memory_table.rs
